Update::leftTime 코드가 길어져서 컴포넌트로 분리
